CV_automation/requirements.txt
supabase>=2.10.0
ollama>=0.4.0
pydantic>=2.10.0
playwright>=1.49.0
markdown2>=2.5.0
aiofiles>=24.1.0
nvidia-ml-py>=12.535.161
pyyaml>=6.0.2
streamlit>=1.41.0
python-dotenv>=1.0.1
pytest>=8.3.4
pytest-asyncio>=0.25.2
responses>=0.25.6
numpy>=1.26.0
pandas>=2.2.0
